+++ tests/testthat/test_prob_remove.R
@@ -1,0 +1,6 @@
+test_that("number of entries behaves as expected", {
+  x = 1:10
+  expect_lt(length(prob_remove(x, 0.5)), length(x))
+  expect_equal(length(prob_remove(x, 1)), length(x))
+  expect_equal(length(prob_remove(x, 0)), 0)
+})