+++ NEWS.md
@@ -1,5 +1,7 @@
 # StratPal (development version)
 
+* introduce S3 class `pre_paleoTS` with simulation procedures, functions to reduce to `paleoTS` via `reduce_to_paleoTS`, and implementaions of niche models and taphonomy via `apply_niche` and `apply_taphonomy`
+
 * add more tests
 
 * add input checks